requirements.txt
# ── LLM & 오케스트레이션 ──────────────────────────────────
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.8
langchain-community>=0.2.0

# ── 트렌드 검색 ───────────────────────────────────────────
tavily-python>=0.3.3          # Tavily Search API

# ── 이미지 처리 ───────────────────────────────────────────
pillow>=10.0.0
matplotlib>=3.8.0             # 성과 분석 차트

# ── 뉴스 수집 ─────────────────────────────────────────────
feedparser>=6.0.0             # RSS 피드 파싱

# ── 스케줄러 ──────────────────────────────────────────────
apscheduler>=3.10.0           # 매일 자동 실행

# ── 텔레그램 봇 ───────────────────────────────────────────
python-telegram-bot>=20.7     # 폰으로 승인/관리

# ── 웹 대시보드 ───────────────────────────────────────────
flask>=3.0.0                  # 운영 현황 웹 UI

# ── 환경 변수 & 유틸 ─────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.27.0
pyngrok>=7.0.0                # Instagram 업로드용 임시 공개 URL
